Prevent llvm from assembling `mov` involving ax/eax/rax and an absolute address as `movabs`
If you specifically want `movabs`, you (or your compiler) should say so. if you use `movabs` whenever it's possible, you will not be able to reassemble capstone disassembly into the original machine code.

This is not a good way to make this change. This table is auto-generated and you should not edit it directly. It's a hack. But it's a lot easier than trying to generate this table again. Keystone uses an old version of LLVM and trying to work with it just for one change like this was not worth it.